requirements.txt
numpy==1.23.3
pandas==1.4.4
PyYAML==6.0
scikit-learn==1.1.2
scipy==1.9.2
pytest
scikit-optimize
